+++ app/src/main/java/classified/classifiedbuzzmovieselector/controllers/MovieInformationActivity.java
@@ -1,5 +1,6 @@
 package classified.classifiedbuzzmovieselector.controllers;
 
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -10,6 +11,7 @@ import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.View;
+import android.widget.Toast;
 
 import classified.classifiedbuzzmovieselector.R;
 import classified.classifiedbuzzmovieselector.model.MovieManager;
@@ -26,17 +28,8 @@ public class MovieInformationActivity extends AppCompatActivity {
         setContentView(R.layout.activity_movie_information);
     }
 
-    /**
-     * Navigates to the rating activity so a user can rate a movie
-     * @param v The current view the user sees
-     */
-    public void onRatingButtonPressed(View v) {
-        Intent intent = new Intent(this, RatingActivity.class);
-        startActivity(intent);
-    }
-
     public void onCancelMovieInformationButtonPressed(View v) {
-        Log.d("RATING ACTIVITY", "Rating was cancelled.");
+        Log.d("RATING ACTIVITY", "Movie Information was cancelled.");
         RecyclerView recList = (RecyclerView) findViewById(R.id.movieResultList);
         recList.setHasFixedSize(true);
         LinearLayoutManager llm = new LinearLayoutManager(this);
@@ -44,5 +37,29 @@ public class MovieInformationActivity extends AppCompatActivity {
         recList.setLayoutManager(llm);
         MovieAdapter ma = new MovieAdapter(MovieManager.getMovies());
         recList.setAdapter(ma);
+    }
+
+    /**
+     * Creates a rating for the movie
+     * @param v The current view
+     */
+    public void onMovieRatingButtonPressed(View v) {
+        Log.d("RATINGACTIVITY", "Rating button was pressed.");
+        //TODO:
+        //Get rating (int) from user
+        //Get user info
+        //Get movie the rating goes with
+        //Pass information to UserRating to create a rating
+        //Add rating to userRatingManager
+        try {
+
+        } catch (Exception e) {
+            CharSequence message;
+            message = "Rating was unsuccessful. Please try again.";
+            Context context = getApplicationContext();
+            int duration = Toast.LENGTH_SHORT;
+            Toast output = Toast.makeText(context, message, duration);
+            output.show();
+        }
     }
 }
